Implement Subclass type in BulkDetails and ComponentDetails.
As the MatML Subclass class is defined in both the BulkDetails and ComponentDetails classes, the Subclass classes were defined and Subclass and Subclass1 for the BulkDetails and the ComponentDetails respectively. Note that Subclass and SubClass1 are internally defined so they are both Strongtypes as they are not defined externally from a MatML type.