Maintain connectivity when changing Id's #10
Linux build had a bug regarding the memory address of tmp. Change the code to be more explicit. Gave the ability of the compiler to work it out, but the code still takes an address of a temp. This may work because the original data is still within the MatML Data tree.